requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3'
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
deprecated==1.2.13
django-filter==21.1
django-stubs-ext==0.4.0
django-stubs==1.10.1
django==4.0.4
djangorestframework-jsonapi==5.0.0
djangorestframework-stubs==1.5.0
djangorestframework==3.13.1
idna==3.3; python_version >= '3'
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.1
kombu==5.2.4
markupsafe==2.1.1
mypy-extensions==0.4.3
mypy==0.942
packaging==21.3
prompt-toolkit==3.0.29
psycopg2==2.9.3
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
types-markdown==3.3.13
types-pytz==2021.3.6
types-pyyaml==6.0.7
types-requests==2.27.20
types-urllib3==1.26.13
typing-extensions==4.2.0
typing==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
